Update cloudalyze.pro
I added three new fields to the output structure to store the deconvolved major and minor axis sizes and position angle. This should facilitate the computation of more robust cloud aspect ratios. I tested a minimal example and it seemed to produce the correct output for both resolved and unresolved structures.